// src/payload-types.ts
export interface Category {
  id: string
  name: string
  slug: string
  color?: string | null
  parent?: (string | null) | Category
  subcategories?: {
    length: number
    docs?: (string | Category)[]
    hasNextPage?: boolean
    totalDocs?: number
  }
  updatedAt: string
  createdAt: string
}